hack to use alsasrc instead of pulsesrc.
to do this properly would require modifications to the core of
galicaster as only 'pulse' and 'audiotest' are hardcoded as audio srcs.
could add an 'alsa' to the list but should probably be done more
generally.

this hack will be ok until we need to use different hardware that
requires pulse :(